Remove JS use from expandStar
This changes somehow causes more temporary variables
to be used than before. There must be someting in
Assign::compile that does scope mangling.

IMO, the reuse of temp variables will be less of a problem
once source maps are in play. Plus, with a full JS AST, temp
var reuse can probably be done with a regular JS optimizer
instead of internally.